pyproject.toml
[project]
name = "litigationtracker"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "django>=5.2.4",
    "django-allauth>=65.10.0",
    "django-crispy-forms>=2.4",
    "django-impersonate>=1.9.5",
    "django-money>=3.5.4",
    "django-simple-history>=3.10.1",
    "psycopg>=3.2.9",
]
